task-manager/requirements.txt
Flask==2.2.2
azure-servicebus==7.8.1